Fixed signout, and pinging
When you sign out, it now will pop up the sign in modal again which is
good. Added in so that whenever you ping, it posts your location first,
before it was just taking the servers last location, which was coming
back wrong sometimes as the user had moved, and the background location
had not worked.